### 2021-02-02 #47 @fgeoffroy
[Re] Assortative matching and search
Dear Editors,

I would like to request a review for the following replication.

**Original articles:**
- Shimer, R., and Smith, L. (2000). Assortative matching and search. Econometrica, 68(2), 343–369. http://doi.org/10.1111/1468-0262.00112
- Smith, L. (2006). The Marriage Model with Search Frictions. J. Political Econ., 114(6), 1124–1144. https://doi.org/10.1086/510440

**Scientific domain:** Game theory
