### 2021-02-15 #47 @rougier
@eddelbuettel @AurelienNioche Could you review this short submission in econometrics?
