### 2021-02-15 #47 @eddelbuettel
I could take a look but the self-selected field is game theory which is less my beef.
