### 2021-02-16 #47 @AurelienNioche
I agree that the label 'game theory' would be a better match than CS. I am also not familiar with Becker's model, but I can check that the code is doing what is supposed to do, and that I can reproduce the results.
