### 2021-02-21 #47 @eddelbuettel
### Reproduction of _Assortative matching and search_  for Shimer and Smith (2000) and Smith (2006)

The submission reproduces results from two seminal paper in game theory and matching.  We find the reproduction to be faithful, replicable, and carefully executed.  Minor comments follow.

#### Paper

The [submitted paper](https://github.com/fgeoffroy/-Re-Assortative-matching-and-search/blob/main/article/article.pdf) is clearly written and introduces the scientific content, problem setting, and relevant references in five pages.  It discusses the results carefullu, in particular the ability to reproduce all but one figure.  The lack of reproducibility of the final is reasoned about in some detail revealing contact with the original author, lack of source code for that part of the original study as well some conjecture about possibly parameter misprints and possible approximation.  Four pages present the reproduced figures which are also presented in the corresponding [reproduction code repository](https://github.com/fgeoffroy/-Re-Assortative-matching-and-search/tree/main/figures).

#### Code

Usage and requirement of the [submitted code](https://github.com/fgeoffroy/-Re-Assortative-matching-and-search/tree/main/code) are detailed appropriately.  We appreciate that the required software packages are both small in numbers (only a Python interpreter plus two packages) and very mainstream (NumPy, Matplotlib). Yet we find the description overly stringent; Python code can be executed in environments other than Conda.  The author lists Ubuntu 18.04 as the platform for his work; we suspect the code would have worked with the default setup. We reproduced everything under the (current) Ubuntu 20.10 release with versions 3.8.6 of Python, 1.18.4 of NumPy and 3.3.0 of Matplotlib. This would suggest that a less restrictive description might be suitable.  If the author were truly concerned about versions, maybe a small Docker container with these versions could be provided too.

The code itself is commented and clearly enough written.  My default editor and it [elpy](https://github.com/jorgenschaefer/elpy) follow some coding guidelines and bemoan a lack of some `def` statements in `main.py` as well as an inappropriate number of lines---but I tend to find those nags to be overly stringent so this is not a real concern.

#### Figures

The main outcome of the reproduction are 20 pdf files.  All appear to be identical to the [submitted versions](https://github.com/fgeoffroy/-Re-Assortative-matching-and-search/tree/main/figures).  As "eye-balling" figure has its limits, we made use of two R packages to i) convert the pdf files to png graphics and to ii) run a "visual test" comparing original and reproduction.  Doing so is difficult, and has limits but we found the package used ([visualTest](https://github.com/MangoTheCat/visualTest), on GitHub but sadly not published to CRAN) appropriate (and we also use it in teaching and marking).  The comparison script, in essence a loop over all all files with temporary file creation, conversion and comparison for each file, is below.

#### Disclaimer

This is my first review for ReScience, so please do lot me know if anything is insufficient or unclear.  Game theory is also not something I have really touched since the graduate school days (quite some time ago) but as there are few other economists among the ReScience reviewer, I hope that this will still be of help.

#### Appendix

The ad-hoc R script follows.  The two top-level directory assignment may need local adjustments.

```r

## Compare two sets of pdf files for https://github.com/ReScience/submissions/issues/47
## Dirk Eddelbuettel, Feb 2021, released under GPL 2 or later

library(pdftools)		# on CRAN: to convert pdf to png
library(visualTest)		# at github.com/MangoTheCat/visualTest: to assess figure similarity

dirA <- "code/figures"          # soft link into author repo
dirB <- "re-run-figures"        # output of reproduction

pdffiles <- list.files(dirA, pattern="*.pdf")

for (f in pdffiles) {
    ## construct filenames in original and reproduced directories
    fileA <- file.path(dirA, f)
    fileB <- file.path(dirB, f)
    stopifnot(`no reproduced file found` = file.exists(fileB))

    ## create two tempfiles and produces png variants
    tfA <- tempfile(fileext=".png")
    tfB <- tempfile(fileext=".png")
    pdf_convert(fileA, format="png", filenames=tfA, verbose=FALSE)
    pdf_convert(fileB, format="png", filenames=tfB, verbose=FALSE)

    ## compare these png variants for exactness (approximately but still...)
    stopifnot(`disparity found` = isSimilar(tfA, tfB, exact=TRUE))

    ## report progress
    cat("Processed and compared", f, "without issues.\n")
}

cat("Done. No issues.\n")
```
