### 2021-02-22 #47 @rougier
@eddelbuettel Many thanks for the (fast) review that is exactly what is expected.

@fgeoffroy Could you address @eddelbuettel comments. I think the inclusion of the R script that @eddelbuettel indly offered would be a nice addition to your work. As for the less restrictive environment, I think it makes sense to loosen it a bit since your work is generic enough.
