+++ code/main.py
@@ -10,7 +10,9 @@ print("FIGURE 1")
 delta = 1
 rho = 100
 r = 1
-production_function = lambda x,y : x * y
+def production_function(x, y): return x * y
+
+
 nList = [10, 50, 100, 500]
 for i in range(len(nList)):
     n = nList[i]
@@ -18,12 +20,8 @@ for i in range(len(nList)):
     find_equilibrium_TU(n, delta, rho, r, production_function, figName)
 
 
-
-
 # From here, all the figures (except fig. 2B) are built with n = 500
 n = 500
-
-
 
 
 # FIGURE 2
@@ -32,7 +30,9 @@ print("FIGURE 2")
 delta = 1
 rho = 100
 r = 1
-production_function = lambda x,y : (x + y - 1) ** 2
+def production_function(x, y): return (x + y - 1) ** 2
+
+
 figName = "fig2_1.pdf"
 find_equilibrium_TU(n, delta, rho, r, production_function, figName)
 
@@ -42,7 +42,9 @@ nFig2B = 501
 delta = 1
 rho = 35
 r = 1
-production_function = lambda x,y : (x + y) ** 2
+def production_function(x, y): return (x + y) ** 2
+
+
 figName = "fig2_2.pdf"
 find_equilibrium_TU(nFig2B, delta, rho, r, production_function, figName)
 
@@ -51,7 +53,9 @@ find_equilibrium_TU(nFig2B, delta, rho, r, production_function, figName)
 delta = 1
 rho = 750
 r = 1
-production_function = lambda x,y : x + y + x * y
+def production_function(x, y): return x + y + x * y
+
+
 figName = "fig2_3.pdf"
 find_equilibrium_TU(n, delta, rho, r, production_function, figName)
 
@@ -60,12 +64,11 @@ find_equilibrium_TU(n, delta, rho, r, production_function, figName)
 delta = 0.5
 rho = 50
 r = 1
-production_function = lambda x,y : x + y + x * y
+def production_function(x, y): return x + y + x * y
+
+
 figName = "fig2_4.pdf"
 find_equilibrium_TU(n, delta, rho, r, production_function, figName)
-
-
-
 
 
 # NON-TRANSFERABLE UTILITY (NTU)
@@ -77,7 +80,9 @@ print("FIGURE 3")
 delta = 0.1
 rho = 30
 r = 0.3
-production_function = lambda x,y : math.exp(x * y)
+def production_function(x, y): return math.exp(x * y)
+
+
 figName = "fig3_1.pdf"
 find_equilibrium_NTU(n, delta, rho, r, production_function, figName)
 
@@ -86,7 +91,9 @@ find_equilibrium_NTU(n, delta, rho, r, production_function, figName)
 delta = 1.1
 rho = 30
 r = 0.3
-production_function = lambda x,y : math.exp(x * y)
+def production_function(x, y): return math.exp(x * y)
+
+
 figName = "fig3_2.pdf"
 find_equilibrium_NTU(n, delta, rho, r, production_function, figName)
 
@@ -95,12 +102,11 @@ find_equilibrium_NTU(n, delta, rho, r, production_function, figName)
 delta = 0.1
 rho = 3
 r = 0.3
-production_function = lambda x,y : x + y + x * y
+def production_function(x, y): return x + y + x * y
+
+
 figName = "fig3_3.pdf"
 find_equilibrium_NTU(n, delta, rho, r, production_function, figName)
-
-
-
 
 
 # NORMAL DISTIBUTION OF TYPES
@@ -110,7 +116,8 @@ print("FIGURE 4")
 delta = 1
 rho = 1000
 r = 1
[…]
