+++ article/header.tex
@@ -98,7 +98,7 @@
 \begin{flushleft}
 \begin{itemize}
 	\renewcommand{\labelitemi}{$\rightarrow$}
-    \item {\href{\replicationURL}{\small \sffamily \replicationCITE}} {\href{http://doi.org/\replicationDOI}{\small \sffamily http://doi.org/\replicationDOI}}
+    \item {\href{\replicationURL}{\small \sffamily \replicationCITE}} {\href{http://doi.org/\replicationDOI}{\small \sffamily http://doi.org/\replicationDOI}} 
     \item {\href{https://www.ssc.wisc.edu/~lones/wp-content/uploads/2016/12/marriage-JPE.pdf}{\small \sffamily Smith, L. (2006). The Marriage Model with Search Frictions. J. Political Econ., 114(6), 1124–1144.}} {\href{https://doi.org/10.1086/510440}{\small \sffamily https://doi.org/10.1086/510440}}
   \end{itemize}\par
 \end{flushleft}

[…]
