### 2021-02-23 #47 @fgeoffroy
Dear all,

Thank you @eddelbuettel for your review and @rougier for your suggestions. I have updated the code, article and README file accordingly in the repository.

To answer @eddelbuettel's comments:

- I agree that the versioning description was too stringent, given that the packages used are very common. I have modified the README file so that the versioning aspect is no longer presented as a major issue. I have kept the Conda environment only for tractability reasons.

- In the `main.py` file, I assume that the lack of `def` statements that you mentioned corresponds to the use of `lambda` functions. Accordingly, I have replaced them for `def` statements in all python files.

- I have not changed the length of the `main.py` file, because I also do not think it is a real concern and, here, it is mostly designed to give information about each figure.

- I have also made minor aesthetic changes so that the python files follow the [PEP-8](https://www.python.org/dev/peps/pep-0008/) guidelines.

- I find the visual comparison idea very relevant, so I have included the R script provided by @eddelbuettel, as suggested by @rougier. Do you agree with this decision @eddelbuettel? It is mentioned that you are the author of this script in the README file, the article and in the script itself. To facilitate the comparison process, I have changed the folder structure: the article figures are now located in `article/figures/` and the output of the main python script goes to `code/figures/`. The R script is named `compare_images.R` and is located in the `code/` subfolder.

Lastly, thanks to @AurelienNioche's reply, I have realized that I had not mentioned the link between the replicated work and Becker's famous model. Thus, I have added a sentence and three relevant references in the Introduction.
