### 2021-03-05 #47 @AurelienNioche
@fgeoffroy
Dear Author,

I found the paper well written, clearly explaining what could be reproduced, what could not (at least not without modifying the parameterization), and how to reproduce the results from Shimer and Smith (2020), and Smith (2006).

Here are a few minor comments, for both the paper and the code:

## Paper

### Results

Figure 3: It may be useful to have a 2 x 2 figure, with an additional panel showing that using delta = 1.1 leads to failure to replicate Figure 2 from [11].

### Not cristal clear
* line 23: it will probably make sense to introduce the articles chronologically
* line 23: "on both types" => this is confusing, as both NTU and TU depend on types of both agents, but not in the same way (3rd line of the equation following line 62 and 3rd line of the equation line 67). Is there a way to make that more clear?
* line 25: this case is called 'Transferable Utility', right?
* line 95: it should be "has been modified by more than epsilon", or similar
* line 127-128: "the higher oneʼs type, the higher the minimum type that one is willing to accept as a partner" => could probably be expressed in a clearer way
* line 129: "Here, there is no Positive Assortative Matching at the equilibrium" => by 'here', you mean "contrarily to the results presented in Fig. 1", right?
* line 160: could you describe here if the results match your expectations? it seems to be, but it would be better to have it said.

### Typos
* line 15: "itterature" instead of "litterature"
* line 47: "in" instead of "is "
* figure 3 caption: For B, delta = 1.1
* several places: "figure X" indead of "Figure X" (capitalization)

## Code

### Execution

Everything worked out of the box for me.

### Redaction

I appreciate the inline comments and spaced code, as well as the very clear and detailed README.

A few minor comments:
* camel case and snake case are both used indifferently for naming variables, while according to the PEP-8 convention, snake case should be preferred
* run_TU.py:
	* line 28: variable "values" is not used
	* line 50: useless parenthesis
	* line 47/48: parameters should be arguments of the function
* run_NTU.py:
	* same as run_TU.py
