+++ code/run_NTU.py
@@ -5,28 +5,27 @@ from matplotlib.ticker import FormatStrFormatter
 import sys
 
 
-def find_equilibrium_NTU(n, delta, rho, r, production_function, figName, distribution="uniform", mu=0.5, sigma=0.1):
+def find_equilibrium_NTU(n, delta, rho, r, production_function, fig_name, tol=1e-12, distribution="uniform", mu=0.5, sigma=0.1):
     # market friction parameter
     psy = (r + delta) / rho
 
     # The distribution of types is discretized to perform a midpoint Riemann sum with n subdivisions
     contributions = np.linspace(1/n/2, 1-1/n/2, n)
     if distribution == "uniform":
-        lDensity = 1
+        l_density = 1
     elif distribution == "normal":
         # Normal distribution
-        lDensity = np.exp(- ((contributions - mu) ** 2) / (2 * sigma ** 2)) / \
+        l_density = np.exp(- ((contributions - mu) ** 2) / (2 * sigma ** 2)) / \
             (sigma * math.sqrt(2 * math.pi))
         # Truncated between 0 and 1
-        lDensity /= (math.erf((1 - mu) / (sigma * math.sqrt(2))) -
+        l_density /= (math.erf((1 - mu) / (sigma * math.sqrt(2))) -
                      math.erf((- mu) / (sigma * math.sqrt(2)))) / 2
     else:
         sys.exit("Warning: The distribution of type should be 'uniform' or 'normal'")
 
     # Initial values
     alphas = np.ones((n, n))
-    values = np.repeat(0., n)
-    uDensity = np.repeat(0., n)
+    u_density = np.repeat(0., n)
 
     # Computing the payoffs for all the types combinations
     payoffs = np.empty([n, n])
@@ -36,55 +35,54 @@ def find_equilibrium_NTU(n, delta, rho, r, production_function, figName, distrib
             y = contributions[j]
             payoffs[i, j] = production_function(x, y)
 
-    listAllAlphas = [alphas.tolist()]
-    keepIterating = True
+    list_all_alphas = [alphas.tolist()]
+    keep_iterating = True
 
     # Main loop
-    while keepIterating:
+    while keep_iterating:
 
         # Updating the unemployment density. Equation (1)
         # """ Fixed point algorithm """
-        tol = 1e-12
-        e = 1
-        uPrev = uDensity
-        while(e > tol):
-            uDensity = delta * lDensity / \
-                (delta + rho * np.dot(alphas, uPrev) / n)  # fixed point iteration
-            e = np.linalg.norm(uPrev - uDensity)
-            uPrev = uDensity
+        e = sys.float_info.max
+        u_prev = u_density
+        while e > tol:
+            u_density = delta * l_density / \
+                (delta + rho * np.dot(alphas, u_prev) / n)  # fixed point iteration
+            e = np.linalg.norm(u_prev - u_density)
+            u_prev = u_density
 
         # Updating the values. Equation (2b)
-        values = (np.dot(alphas * payoffs, uDensity)) / (n * psy + (np.dot(alphas, uDensity)))
+        values = (np.dot(alphas * payoffs, u_density)) / (n * psy + (np.dot(alphas, u_density)))
 
         # Updating the matching set. Equation (3b)
-        newAlphas = np.zeros([n, n])
+        new_alphas = np.zeros([n, n])
         for i in range(n):
             for j in range(n):
                 if payoffs[i, j] >= values[i] and payoffs[j, i] >= values[j]:
-                    newAlphas[i, j] = 1
+                    new_alphas[i, j] = 1
[…]

[…]
