### 2021-03-09 #47 @fgeoffroy
Dear @AurelienNioche,

Thank you very much for your review and suggestions.

To answer your comments on the article:

- For Figure 3, I have added the sub-figure which shows a failure to replicate Figure 2 from [11] with delta = 1.1.

- In the third paragraph of the Introduction, I have modified the description of TU and NTU to make the distinction clearer.

- In the last step of the algorithm, the matching set `A` is a matrix of zeros and ones. The stopping criteria of the algorithm is thus simply when the matrix is identical to its value in the previous step.

- In the "Shimer & Smith 2000" section, I have modified the description of PAM to improve clarity as you suggested.

- In the "Gaussian distribution" section, I have added a description of the results, especially on the effect of the standard deviation.

To answer your comments on the code:

- I have modified the code to use the snake case naming convention.

- The useless variable and parenthesis have been removed.

- The parameter `tol` (the absolute tolerance level for the fixed-point iteration) is now an argument of the `find_equilibrium_TU` and `find_equilibrium_NTU` functions. The variable `e` is the error which is minimized by the fixed-point iteration. Its initial value was previously 1. For clarity, I have changed it to `sys.float_info.max`, the maximal float value in Python.

Thank you again @eddelbuettel and @AurelienNioche for you comments, I think they really improved the readability and quality of this replication. Thank you @rougier for the very fast and efficient review!
