### 2021-03-15 #47 @rougier
@AurelienNioche If you're satisfied with @fgeoffroy, I think we can accept the submission! Congratulations @fgeoffroy !
I'll need the URL to your article repository for the publication. In the meantime, can you save your code at https://www.softwareheritage.org/save-and-reference-research-software/ (the process is straightforwared and you should obtain a swid that is needed for publication).
