+++ article/metadata.tex
@@ -1,25 +1,25 @@
 % DO NOT EDIT - automatically generated from metadata.yaml
 
-\def \codeURL{https://github.com/rescience-c/template}
+\def \codeURL{https://github.com/fgeoffroy/-Re-Assortative-matching-and-search}
 \def \codeDOI{}
 \def \codeSWH{}
 \def \dataURL{}
 \def \dataDOI{}
-\def \editorNAME{}
-\def \editorORCID{}
-\def \reviewerINAME{}
-\def \reviewerIORCID{}
-\def \reviewerIINAME{}
-\def \reviewerIIORCID{}
+\def \editorNAME{Nicolas P. Rougier}
+\def \editorORCID{0000-0002-6972-589X}
+\def \reviewerINAME{Dirk Eddelbuettel}
+\def \reviewerIORCID{0000-0001-6419-907X}
+\def \reviewerIINAME{Aurélien Nioche}
+\def \reviewerIIORCID{0000-0002-0567-2637}
 \def \dateRECEIVED{02 February 2021}
-\def \dateACCEPTED{}
-\def \datePUBLISHED{}
+\def \dateACCEPTED{15 March 2021}
+\def \datePUBLISHED{15 March 2021}
 \def \articleTITLE{[Re] Assortative matching and search}
 \def \articleTYPE{Replication}
 \def \articleDOMAIN{Game Theory}
 \def \articleBIBLIOGRAPHY{bibliography.bib}
 \def \articleYEAR{2021}
-\def \reviewURL{}
+\def \reviewURL{https://github.com/ReScience/submissions/issues/47}
 \def \articleABSTRACT{}
 \def \replicationCITE{Shimer, R., and Smith, L. (2000). Assortative matching and search. Econometrica, 68(2), 343–369.}
 \def \replicationBIB{shimer_assortative_2000}
