### 2021-03-16 #47 @fgeoffroy
@rougier I have saved the code at the following SWHID: swh:1:dir:c4b7d774ab458f90c433e321ab969d3f12e410de
I have also reserved a DOI (10.5281/zenodo.4607228), but I just miss some information before I publish it:
- the article number (asked in the `metadata.yaml` file)
- the journal volume
- the journal issue

Or should I proceed with the publication on Zenodo and you give me these information afterwards? I'm not sure about the procedure.
