### 2021-03-24 #47 @rougier
Here is the sandbox version: https://sandbox.zenodo.org/record/753083. Can you check if this looks good? If so, I'll publish the final article.
