### 2021-03-24 #47 @fgeoffroy
Thank you very much for handling this!
The long SWHID is somehow messing with the format, but it's not problematic and everything else seems perfect.
