### 2021-03-24 #47 @rougier
I noticed that too yes. I think we may need to change the template a bit. If you want to fix your article, I can make a PR on your repo and you can try to fix it. Then I'll publish it.
