### 2021-03-24 #47 @fgeoffroy
I updated the article on the github repo with a fix for the SWHID, the article number, journal volume and issue and the sandbox DOI that you created. Is that enough? If it is to much trouble, then you can just publish the previous version, the formatting problem is really minor.
