### 2021-03-29 #47 @rougier
Sorry for the delay. Can you tell what did you change in the template such I can do the same on my side.  The final (real) DOI will be different such that I need to be able to compile your article.
