### 2021-03-29 #47 @fgeoffroy
Yes sorry, I simply added a linebreak before the SWH in the header.tex file:
`\ifdefempty{\codeSWH}{.}{ -- \\ SWH \href{https://archive.softwareheritage.org/\codeSWH/}{\detokenize\expandafter{\codeSWH}}.}`
