+++ article/metadata.tex
@@ -2,7 +2,7 @@
 
 \def \codeURL{https://github.com/fgeoffroy/-Re-Assortative-matching-and-search}
 \def \codeDOI{}
-\def \codeSWH{}
+\def \codeSWH{swh:1:dir:c4b7d774ab458f90c433e321ab969d3f12e410de}
 \def \dataURL{}
 \def \dataDOI{}
 \def \editorNAME{Nicolas P. Rougier}
@@ -13,7 +13,7 @@
 \def \reviewerIIORCID{0000-0002-0567-2637}
 \def \dateRECEIVED{02 February 2021}
 \def \dateACCEPTED{15 March 2021}
-\def \datePUBLISHED{15 March 2021}
+\def \datePUBLISHED{24 March 2021}
 \def \articleTITLE{[Re] Assortative matching and search}
 \def \articleTYPE{Replication}
 \def \articleDOMAIN{Game Theory}
@@ -29,10 +29,10 @@
 \def \contactEMAIL{felix.geoffroy.fr@gmail.com}
 \def \articleKEYWORDS{Search-and-matching, markets, biological markets, game theory, Python}
 \def \journalNAME{ReScience C}
-\def \journalVOLUME{4}
+\def \journalVOLUME{7}
 \def \journalISSUE{1}
-\def \articleNUMBER{}
-\def \articleDOI{}
+\def \articleNUMBER{2}
+\def \articleDOI{10.5281/zenodo.4646680}
 \def \authorsFULL{Félix Geoffroy}
 \def \authorsABBRV{F. Geoffroy}
 \def \authorsSHORT{Geoffroy}
