### 2021-03-30 #47 @rougier
Thanks. It's online: https://rescience.github.io/bibliography/Geoffroy_2021.html
